Use fp16 compute for 4-bit GRPO

+++ scripts/train_llm_grpo_all_stages.py
@@ -108,7 +108,7 @@ def load_model(model_name: str, use_unsloth: bool) -> tuple[Any, Any]:
         quantization_config = BitsAndBytesConfig(
             load_in_4bit=True,
             bnb_4bit_quant_type="nf4",
-            bnb_4bit_compute_dtype=torch.bfloat16,
+            bnb_4bit_compute_dtype=torch.float16,
             bnb_4bit_use_double_quant=True,
         )
         model = AutoModelForCausalLM.from_pretrained(
